refactor: moved getFileMetadata to file handler
